state.tsにRepoとGithubStateを移動
